Move onewire cli to a separate app
