+++ applications/main/onewire/application.fam
@@ -1,0 +1,14 @@
+App(
+    appid="onewire",
+    name="1-Wire",
+    apptype=FlipperAppType.METAPACKAGE,
+    provides=["onewire_start"],
+)
+
+App(
+    appid="onewire_start",
+    apptype=FlipperAppType.STARTUP,
+    entry_point="onewire_on_system_start",
+    requires=["onewire"],
+    order=60,
+)
